Fix missing imports; move additional functions to utils
